+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -54,7 +54,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 - Added support for listing all Event Grid topics in a subscription via the command `azmcp_eventgrid_topic_list`. [[#43](https://github.com/microsoft/mcp/pull/43)]
 - Added support for retrieving knowledge index schema information in Azure AI Foundry projects via the command `azmcp_foundry_knowledge_index_schema`. [[#41](https://github.com/microsoft/mcp/pull/41)]
 - Added the following Azure Managed Lustre commands: [[#110](https://github.com/microsoft/mcp/issues/110)]
-  - `azmcp_azuremanagedlustre_filesystem_subnet-size-validate`: Check if the subnet can host the target Azure Managed Lustre SKU and size.
+  - `azmcp_azuremanagedlustre_filesystem_subnetsize_validate`: Check if the subnet can host the target Azure Managed Lustre SKU and size.
 - Added support for listing service health events in a subscription via the command `azmcp_resourcehealth_service-health-events_list`. [[#367](https://github.com/microsoft/mcp/pull/367)]
 
 ### Breaking Changes
@@ -81,6 +81,9 @@ The Azure MCP Server updates automatically by default whenever a new release com
   - Microsoft.Azure.Cosmos.Aot: `0.1.1-preview.1` → `0.1.2-preview.1` [[#383](https://github.com/microsoft/mcp/pull/383)]
 - Updated the following dependency to improve .NET Ahead-of-Time (AOT) compilation support: [[#363](https://github.com/microsoft/mcp/pull/363)]
   - Azure.ResourceManager.StorageCache: `1.3.1` → `1.3.2`
+
+- Changed the following Azure Managed Lustre commands:
+  - `azmcp_azuremanagedlustre_filesystem_required-subnet-size` becomes `azmcp_azuremanagedlustre_filesystem_subnetsize_ask`
 
 ## 0.5.12 (2025-09-04)